python_pro/requirements.txt
pillow==10.4.0
requests==2.32.3
lxml==5.3.0
mypy==1.11.2
pylint==3.3.1
types-requests==2.32.0.20240914
pytest==8.3.3
certifi==2024.8.30
pymongo==4.10.1
types-certifi==2021.10.8.3
redis==4.3.4
cql==1.4.0
cassandra-driver==3.29.2
aiohttp==3.10.10
aiofiles==24.1.0
django-bootstrap-v5==1.0.11